Hacked together a fix to the crash when exiting the program.
Problem is likely that things are deleted in the wrong order. The proper way to fix this would be to make loading and unloading of assets explicit instead of using RAII. This would be a pretty big change though and it might not worth it this late in development. Need to think about it some more.